backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
mysqlclient==2.2.4
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
redis==5.0.8
celery==5.4.0
boto3==1.35.0
django-storages==1.14.4
phonenumbers==8.13.45
Pillow==10.4.0
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.7.0
python-decouple==3.8
